Refs #31349. Move cyclic workspace selector to BasicWidget.
