+++ scripts/Muon/GUI/Common/fitting_widgets/basic_fitting/basic_fitting_view.py
@@ -10,6 +10,7 @@ from mantidqt.utils.qt import load_ui
 
 from Muon.GUI.Common.fitting_widgets.basic_fitting.fit_controls_view import FitControlsView
 from Muon.GUI.Common.fitting_widgets.basic_fitting.fit_function_options_view import FitFunctionOptionsView
+from Muon.GUI.Common.fitting_widgets.basic_fitting.workspace_selector_view import WorkspaceSelectorView
 from Muon.GUI.Common.message_box import warning
 
 from qtpy.QtWidgets import QWidget
@@ -28,9 +29,11 @@ class BasicFittingView(QWidget, ui_fitting_layout):
         self.setupUi(self)
 
         self.fit_controls = FitControlsView(self)
+        self.workspace_selector = WorkspaceSelectorView(self)
         self.fit_function_options = FitFunctionOptionsView(self, is_frequency_domain)
 
         self.fit_controls_layout.addWidget(self.fit_controls)
+        self.workspace_selector_layout.addWidget(self.workspace_selector)
         self.fit_function_options_layout.addWidget(self.fit_function_options)
 
         self.disable_tab_observer = GenericObserver(self.disable_view)
@@ -53,6 +56,10 @@ class BasicFittingView(QWidget, ui_fitting_layout):
     def set_slot_for_plot_guess_changed(self, slot) -> None:
         """Connect the slot for the Plot Guess checkbox."""
         self.fit_controls.set_slot_for_plot_guess_changed(slot)
+
+    def set_slot_for_dataset_changed(self, slot) -> None:
+        """Connect the slot for the display workspace combo box being changed."""
+        self.workspace_selector.set_slot_for_dataset_changed(slot)
 
     def set_slot_for_fit_name_changed(self, slot) -> None:
         """Connect the slot for the fit name being changed by the user."""
@@ -86,6 +93,10 @@ class BasicFittingView(QWidget, ui_fitting_layout):
         """Connect the slot for the Use raw option."""
         self.fit_function_options.set_slot_for_use_raw_changed(slot)
 
+    def set_workspace_combo_box_label(self, text: str) -> None:
+        """Sets the label text next to the workspace selector combobox."""
+        self.workspace_selector.set_workspace_combo_box_label(text)
+
     def set_datasets_in_function_browser(self, dataset_names: list) -> None:
         """Sets the datasets stored in the FunctionBrowser."""
         self.fit_function_options.set_datasets_in_function_browser(dataset_names)
@@ -94,6 +105,10 @@ class BasicFittingView(QWidget, ui_fitting_layout):
         """Sets the index of the current dataset."""
         if dataset_index is not None:
             self.fit_function_options.set_current_dataset_index(dataset_index)
+
+    def update_dataset_name_combo_box(self, dataset_names: list) -> None:
+        """Update the data in the parameter display combo box."""
+        self.workspace_selector.update_dataset_name_combo_box(dataset_names)
 
     def update_local_fit_status_and_chi_squared(self, fit_status: str, chi_squared: float) -> None:
         """Updates the view to show the status and results from a fit."""
@@ -109,6 +124,25 @@ class BasicFittingView(QWidget, ui_fitting_layout):
     def update_fit_function(self, fit_function: IFunction) -> None:
         """Updates the parameters of a fit function shown in the view."""
         self.fit_function_options.update_function_browser_parameters(False, fit_function)
+
+    @property
+    def current_dataset_name(self) -> str:
+        """Returns the selected dataset name."""
+        return self.workspace_selector.current_dataset_name
+
+    @current_dataset_name.setter
+    def current_dataset_name(self, dataset_name: str) -> None:
+        """Sets the currently selected dataset name."""
+        self.workspace_selector.current_dataset_name = dataset_name
+
+    def number_of_datasets(self) -> int:
+        """Returns the number of dataset names loaded into the widget."""
+        return self.workspace_selector.number_of_datasets()
+
+    @property
+    def current_dataset_index(self) -> str:
+        """Returns the index of the currently displayed dataset."""
+        return self.workspace_selector.current_dataset_index
 
     @property
     def fit_object(self) -> IFunction:
@@ -178,10 +212,6 @@ class BasicFittingView(QWidget, ui_fitting_layout):
     def function_name(self, function_name: str) -> None:
         """Sets the function name being used."""
         self.fit_function_options.function_name = function_name
-
-    def number_of_datasets(self) -> int:
-        """Returns the number of dataset names loaded into the widget."""
-        return self.fit_function_options.number_of_datasets()
 
     def warning_popup(self, message: str) -> None:
         """Displays a warning message."""
