+++ scripts/Muon/GUI/Common/fitting_widgets/general_fitting/general_fitting_options_view.py
@@ -11,8 +11,6 @@ from qtpy.QtWidgets import QWidget
 ui_general_fitting_options, _ = load_ui(__file__, "general_fitting_options.ui")
 
 MA_FIT_BY_OPTIONS = ["Run", "Group/Pair"]
-SINGLE_FIT_LABEL = "Select Workspace"
-SIMULTANEOUS_FIT_LABEL = "Display parameters for"
 
 
 class GeneralFittingOptionsView(QWidget, ui_general_fitting_options):
@@ -25,19 +23,12 @@ class GeneralFittingOptionsView(QWidget, ui_general_fitting_options):
         super(GeneralFittingOptionsView, self).__init__(parent)
         self.setupUi(self)
 
-        self.increment_parameter_display_button.clicked.connect(self.increment_dataset_name_combo_box)
-        self.decrement_parameter_display_button.clicked.connect(self.decrement_dataset_name_combo_box)
-
         self.disable_simultaneous_fit_options()
 
         if is_frequency_domain:
             self.hide_simultaneous_fit_options()
         else:
             self._setup_simultaneous_fit_by_combo_box(MA_FIT_BY_OPTIONS)
-
-    def set_slot_for_dataset_changed(self, slot) -> None:
-        """Connect the slot for the display workspace combo box being changed."""
-        self.dataset_name_combo_box.currentIndexChanged.connect(slot)
 
     def set_slot_for_fitting_mode_changed(self, slot) -> None:
         """Connect the slot for the simultaneous fit check box."""
@@ -51,66 +42,10 @@ class GeneralFittingOptionsView(QWidget, ui_general_fitting_options):
         """Connect the slot for the fit specifier combo box being changed."""
         self.simul_fit_by_specifier.currentIndexChanged.connect(slot)
 
-    def update_dataset_name_combo_box(self, dataset_names: list) -> None:
-        """Update the data in the parameter display combo box."""
-        old_name = self.dataset_name_combo_box.currentText()
-
-        self.update_dataset_names_combo_box(dataset_names)
-
-        new_index = self.dataset_name_combo_box.findText(old_name)
-        new_index = new_index if new_index != -1 else 0
-
-        self.dataset_name_combo_box.setCurrentIndex(new_index)
-        # This signal isn't always sent, so I will emit it manually.
-        self.dataset_name_combo_box.currentIndexChanged.emit(new_index)
-
-    def update_dataset_names_combo_box(self, dataset_names: list) -> None:
-        """Update the datasets displayed in the dataset name combobox."""
-        self.dataset_name_combo_box.blockSignals(True)
-        self.dataset_name_combo_box.clear()
-        self.dataset_name_combo_box.addItems(dataset_names)
-        self.dataset_name_combo_box.blockSignals(False)
-
-    def increment_dataset_name_combo_box(self) -> None:
-        """Increment the parameter display combo box."""
-        index = self.dataset_name_combo_box.currentIndex()
-        count = self.dataset_name_combo_box.count()
-
-        if index < count - 1:
-            self.dataset_name_combo_box.setCurrentIndex(index + 1)
-        else:
-            self.dataset_name_combo_box.setCurrentIndex(0)
-
-    def decrement_dataset_name_combo_box(self) -> None:
-        """Decrement the parameter display combo box."""
-        index = self.dataset_name_combo_box.currentIndex()
-        count = self.dataset_name_combo_box.count()
-
-        if index != 0:
-            self.dataset_name_combo_box.setCurrentIndex(index - 1)
-        else:
-            self.dataset_name_combo_box.setCurrentIndex(count - 1)
-
     def _setup_simultaneous_fit_by_combo_box(self, fit_by_options: list) -> None:
         """Populate the simultaneous fit by combo box."""
         for item in fit_by_options:
             self.simul_fit_by_combo.addItem(item)
-
-    @property
-    def current_dataset_name(self) -> str:
-        """Returns the selected dataset name."""
-        return str(self.dataset_name_combo_box.currentText())
-
-    @current_dataset_name.setter
-    def current_dataset_name(self, dataset_name: str) -> None:
-        """Sets the currently selected dataset name."""
-        index = self.dataset_name_combo_box.findText(dataset_name)
-        if index != -1:
-            self.dataset_name_combo_box.setCurrentIndex(index)
-
-    def number_of_datasets(self) -> int:
-        """Returns the number of dataset names loaded into the widget."""
-        return self.dataset_name_combo_box.count()
 
     @property
     def simultaneous_fit_by(self) -> str:
@@ -130,22 +65,6 @@ class GeneralFittingOptionsView(QWidget, ui_general_fitting_options):
     def simultaneous_fit_by_specifier(self) -> str:
         """Returns the run, group or pair name."""
         return self.simul_fit_by_specifier.currentText()
-
-    @property
-    def current_dataset_index(self) -> int:
-        """Returns the index of the currently displayed dataset."""
-        current_index = self.dataset_name_combo_box.currentIndex()
-        return current_index if current_index != -1 else None
-
-    def switch_to_simultaneous(self) -> None:
-        """Switches the view to simultaneous fit mode."""
-        self.set_workspace_combo_box_label(SIMULTANEOUS_FIT_LABEL)
-        self.enable_simultaneous_fit_options()
-
-    def switch_to_single(self) -> None:
-        """Switches the view to single fit mode."""
-        self.set_workspace_combo_box_label(SINGLE_FIT_LABEL)
-        self.disable_simultaneous_fit_options()
 
     def hide_simultaneous_fit_options(self) -> None:
         """Hides the simultaneous fit options."""
@@ -174,9 +93,6 @@ class GeneralFittingOptionsView(QWidget, ui_general_fitting_options):
         self.simul_fit_checkbox.blockSignals(True)
         self.simul_fit_checkbox.setChecked(simultaneous)
         self.simul_fit_checkbox.blockSignals(False)
-
-    def set_workspace_combo_box_label(self, text: str) -> None:
-        self.workspace_combo_box_label.setText(text)
 
     def setup_fit_by_specifier(self, choices: list) -> None:
         """Setup the fit by specifier combo box."""
